.FIX the month span and week span, also make percentiles rolling, last 30 days percentile

+++ api/src/main/java/ca/bc/gov/educ/api/penrequest/struct/v1/PenRequestStats.java
@@ -21,14 +21,14 @@ public class PenRequestStats {
    * JAN 20
    * FEB 30 ......
    */
-  Map<String, Integer> completionsInLastTwelveMonth;
+  Map<String, Long> completionsInLastTwelveMonth;
   /**
    * contains number of completed GMP requests in last week from current day.
    * ex:- if current day is Wednesday -> it will show from Last Wednesday to Tuesday.
    * WED 20
    * THURS 30 ......
    */
-  Map<String, Integer> completionsInLastWeek;
+  Map<String, Long> completionsInLastWeek;
 
   /**
    * Numbers for different statuses in last 12 month.
